Modification et Suppression via formulaire - fini

+++ main/src/View/AllBookingsJPanel.java
@@ -22,22 +22,24 @@ public class AllBookingsJPanel extends JPanel {
             this.frameContainer = frameContainer;
             controller = new ApplicationController();
             bookings = controller.getAllBookings();
+            this.setLayout(new BorderLayout());
             AllBookingsModel model = new AllBookingsModel(bookings);
             JTable bookingsTable = new JTable(model);
             bookingsTable.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
             listSelect = bookingsTable.getSelectionModel();
             bookingsTable.setPreferredScrollableViewportSize(new Dimension(900, 100));
 
-            this.add(new JScrollPane(bookingsTable));
+            this.add(new JScrollPane(bookingsTable), BorderLayout.NORTH);
             Panel buttonsPanel = new Panel();
             modification = new JButton("Modification");
             modification.addActionListener(new ModifyListener());
             buttonsPanel.add(modification);
             delete = new JButton("Suppression");
             delete.addActionListener(new DeleteListener());
+
             buttonsPanel.add(delete);
-            this.add(buttonsPanel, BorderLayout.SOUTH);
-            this.setLayout(new FlowLayout());
+            this.add(buttonsPanel, BorderLayout.CENTER);
+
             this.setVisible(true);
         }
         catch (AllBookingsException e) {
@@ -54,14 +56,14 @@ public class AllBookingsJPanel extends JPanel {
         @Override
         public void actionPerformed(ActionEvent e) {
 
-            int indiceLigneSelect = listSelect.getMinSelectionIndex();
+            int indSelectedLine = listSelect.getMinSelectionIndex();
 
-            if(indiceLigneSelect != -1) {
+            if(indSelectedLine != -1) {
 
-                int reponse = JOptionPane.showConfirmDialog(null,"Etes-vous sûr de vouloir modifier ce type d'article ?", "Modification", JOptionPane.YES_NO_OPTION);
+                int reponse = JOptionPane.showConfirmDialog(null,"Etes-vous sûr de vouloir modifier cette réservation ?", "Modification", JOptionPane.YES_NO_OPTION);
 
                 if(reponse == 0) {
-                    Booking booking = bookings.get(indiceLigneSelect);
+                    Booking booking = bookings.get(indSelectedLine);
 
                     try {
                         frameContainer.removeAll();
@@ -84,13 +86,24 @@ public class AllBookingsJPanel extends JPanel {
 
         @Override
         public void actionPerformed(ActionEvent e) {
-            try {
-                int indSelectedLine = listSelect.getMinSelectionIndex();
-                Booking booking = bookings.get(indSelectedLine);
-                controller.deleteBooking(booking);
-                setVisible(true);
-            } catch (DeleteException exception) {
-                JOptionPane.showMessageDialog(null,exception.getMessage());
+            int indSelectedLine = listSelect.getMinSelectionIndex();
+            if (indSelectedLine != -1) {
+                int reponse = JOptionPane.showConfirmDialog(null, "Etes-vous sûr de vouloir supprimer cette réservation ?", "Suppression", JOptionPane.YES_NO_OPTION);
+                if (reponse == 0) {
+                    Booking booking = bookings.get(indSelectedLine);
+                    try {
+                        controller.deleteBooking(booking);
+                        frameContainer.removeAll();
+                        frameContainer.revalidate();
+                        frameContainer.repaint();
+                        frameContainer.add(new AllBookingsJPanel(frameContainer));
+                    } catch (DeleteException exception) {
+                        JOptionPane.showMessageDialog(null, exception.getMessage());
+                    }
+                }
+
+            } else {
+                JOptionPane.showMessageDialog(null, "Sélectionnez une ligne à supprimer");
             }
         }
     }
